perf: big big numbers for states>4 and order=2

+++ app/automata/cellular_automata.go
@@ -10,7 +10,7 @@ type CellularAutomata struct {
 	Order      int
 	Columns    int
 	Rows       int
-	RuleNumber int64
+	RuleNumber float64
 	FirstLine  FirstLine
 }
